Create new shardingsphere-sql-federation-executor-core module to handle common executor logic
